\name{delta.method.covariance.GMM}
\alias{delta.method.covariance.GMM}
\title{Covariance matrix of the parameters using delta method}
\usage{
delta.method.covariance.GMM(covariance, gradient, size)
}
\arguments{
  \item{covariance}{Covariance matrix of the moment
  conditions}

  \item{gradient}{Gradient matrix of the moment conditions}

  \item{size}{Sample size}
}
\value{
The covariance matrix of the parameters
}
\description{
Covariance matrix of the parameters using delta method
}
\author{
François Pelletier
}

