\name{fourmoments.gmm.vector}
\alias{fourmoments.gmm.vector}
\title{GMM vector for 4 first central moments conditions}
\usage{
fourmoments.gmm.vector(param, sample, meanf, variancef, skewnessf, kurtosisf)
}
\arguments{
  \item{param}{Estimated parameters}

  \item{sample}{Data Sample}

  \item{meanf}{Mean function}

  \item{variancef}{Variance function}

  \item{skewnessf}{Skewness function}

  \item{kurtosisf}{Kurtosis function}
}
\value{
A four column matrix of differences
}
\description{
GMM vector for 4 first central moments conditions
}
\author{
François Pelletier
}

