\name{mean.variance.gmm.vector}
\alias{mean.variance.gmm.vector}
\title{GMM vector for mean and variance moment conditions}
\usage{
\method{mean}{variance.gmm.vector}(param, sample, meanf, variancef)
}
\arguments{
  \item{param}{Estimated parameters}

  \item{sample}{Data Sample}

  \item{meanf}{Mean function}

  \item{variancef}{Variance function}
}
\value{
A two column matrix of differences
}
\description{
GMM vector for mean and variance moment conditions
}
\author{
François Pelletier
}

