\name{meanvariance.gmm.vector}
\alias{meanvariance.gmm.vector}
\title{GMM vector for mean and variance moment conditions}
\usage{
meanvariance.gmm.vector(param, data, meanf, variancef)
}
\arguments{
  \item{param}{Estimated parameters}

  \item{data}{Data Sample}

  \item{meanf}{Mean function}

  \item{variancef}{Variance function}
}
\value{
A two column matrix of differences
}
\description{
GMM vector for mean and variance moment conditions
}
\author{
François Pelletier
}

