\name{obj.gmmGAL.mu}
\alias{obj.gmmGAL.mu}
\title{Objective function for the GMM method}
\usage{
obj.gmmGAL.mu(conditions.vector, ..., W = diag(length(conditions.vector)))
}
\arguments{
  \item{conditions.vector}{Vector of moment conditions}

  \item{...}{Parameters of the vector of moment conditions}

  \item{W}{Weighting matrix}
}
\value{
A scalar value
}
\description{
Objective function for the GMM method
}
\author{
François Pelletier
}

