\name{covariance.GMM}
\alias{covariance.GMM}
\title{Estimated covariance matrix}
\usage{
covariance.GMM(conditions.vector, param, sample...)
}
\arguments{
  \item{conditions.vector}{Vector of moment conditions}

  \item{param}{Vector of estimated parameters}

  \item{sample}{Individual data sample}

  \item{...}{Functions of the vector of moment conditions}
}
\value{
A square covariance matrix
}
\description{
Estimated covariance matrix
}
\author{
François Pelletier
}

