\name{covariance.GMM}
\alias{covariance.GMM}
\title{Estimated covariance matrix}
\usage{
covariance.GMM(conditions.vector, n, ...)
}
\arguments{
  \item{conditions.vector}{Vector of moment conditions}

  \item{n}{Sample size}

  \item{...}{Parameters of the vector of moment conditions}
}
\value{
A square covariance matrix
}
\description{
Estimated covariance matrix
}
\author{
François Pelletier
}

