\name{objective.GMM}
\alias{objective.GMM}
\title{Objective function for the GMM method}
\usage{
objective.GMM(param, conditions.vector, data, W, ...)
}
\arguments{
  \item{param}{Vector of parameters}

  \item{conditions.vector}{Vector of moment conditions}

  \item{data}{Individual data sample}

  \item{W}{Weighting matrix}

  \item{...}{Functions of the vector of moment conditions}
}
\value{
A scalar value
}
\description{
Objective function for the GMM method
}
\author{
François Pelletier
}

