\name{optim.GMM}
\alias{optim.GMM}
\title{Optimization for GMM method}
\usage{
optim.GMM(start, conditions.vector, data, W, ...)
}
\arguments{
  \item{start}{Starting values for the parameters}

  \item{conditions.vector}{Vector of moment conditions}

  \item{data}{Individual data sample}

  \item{W}{Weighting matrix}

  \item{...}{Functions of the vector of moment conditions}
}
\value{
a list with optimization results
}
\description{
Optimization for GMM method
}
\author{
François Pelletier
}

